refactor(backend): change types to entity
