--- apps/backend/src/types/location.ts
@@ -1,10 +1,0 @@
-import { z } from 'zod'
-
-export const locationSchema = z
-  .object({
-    lat: z.number().min(-90).max(90),
-    lon: z.number().min(-180).max(180),
-  })
-  .strict()
-
-export type Location = z.infer<typeof locationSchema>
